requirements.txt
fastapi
uvicorn
torch
ultralytics
tensorflow
pillow
opencv-python
pymongo
python-dotenv
